refactor: move PodcastCategorySection out of feature components folder
